Rename main.py to wiki_sync.py
